Fix handling of free buddy merging in PMM
When merging a block with its buddy, the PMM forgot to account for the
fact that the block was already on a free list for its old order, and
would (potentially) place the same block on a free list for a different
order. This could cause very hard-to-debug issues, especially because
the heap uses the same part of the page used by the PMM's free list for
its own intrusive linked list.

This meant that the issue could (and did, at least with the current
QEMU/OVMF versions) manifest itself as weird heap corruption issues,
including partial heap slabs that were pointed to by the heap's linked
list implementation but were ostensibly "unlinked" themselves, because
they had been unlinked by the PMM. Of course, the issue was very
sensitive to the exact arrangement/orders of the physical pages in
question, which can be affected by anything from firmware to the size of
the kernel image.

The solution, of course, is to make sure that the free buddies of a
block being freed are removed from their respective free lists before
being merged into a larger block.